+++ src/App.tsx
@@ -1,11 +1,11 @@
 import './styles/index.css';
 
 import { useState } from 'react';
-import { LOCAL_STORAGE_CARD_KEY } from './constant';
-import { CardInfo, PageInfo } from './types';
 import { useLocalStorage } from './hooks/useLocalStorage';
 import { AddCardPage } from './pages/AddCardPage';
 import { HomePage } from './pages/HomePage';
+import { CardInfo, PageInfo } from './types/types';
+import { LOCAL_STORAGE_CARD_KEY } from './constants/constant';
 
 export default function App() {
   const [page, setPage] = useState<PageInfo>('homePage');
