+++ src/components/addCardPage/CardNumberInput/CardNumberInput.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
+import { TAB_INDEX_INFO } from '../../../constants/constant';
 import { UseInputProps } from '../../../hooks/useInput';
-import { TAB_INDEX_INFO } from '../../../constant';
-import { Input } from '../../common/Input';
 import { Error } from '../../common/Error';
+import { Input } from '../../common/Input';
 
 interface CardNumberInputProps {
   firstNumberInformation: UseInputProps;
