+++ src/components/addCardPage/CvcInput/CvcInput.tsx
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
+import { TAB_INDEX_INFO } from '../../../constants/constant';
 import { UseInputProps } from '../../../hooks/useInput';
-import { TAB_INDEX_INFO } from '../../../constant';
-import { Input } from '../../common/Input';
-import { InformationButton } from '../../common/InformationButton';
-import { ToolTip } from '../../common/ToolTip';
 import { Error } from '../../common/Error';
+import { InformationButton } from '../../common/InformationButton';
+import { Input } from '../../common/Input';
+import { ToolTip } from '../../common/ToolTip';
 
 interface CvcInputProps {
   cvcInformation: UseInputProps;
