+++ src/components/addCardPage/PasswordInput/PasswordInput.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
+import { TAB_INDEX_INFO } from '../../../constants/constant';
 import { UseInputProps } from '../../../hooks/useInput';
-import { TAB_INDEX_INFO } from '../../../constant';
-import { Input } from '../../common/Input';
 import { Error } from '../../common/Error';
+import { Input } from '../../common/Input';
 
 export interface PasswordInputProps {
   firstPasswordInformation: UseInputProps;
