+++ src/components/common/Modal/Modal.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react';
-import Modal from './Modal';
 import { SelectCardCompany } from '../../addCardPage/SelectCardCompany';
+import Modal from './Modal';
 
 const meta: Meta<typeof Modal> = {
   component: Modal,
