+++ src/components/common/ToolTip/ToolTip.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { TAB_INDEX_INFO } from '../../../constant';
+import { TAB_INDEX_INFO } from '../../../constants/constant';
 
 interface ToolTipProps {
   text: string;
