+++ src/pages/AddCardPage/AddCardPage.tsx
@@ -1,31 +1,31 @@
 import React, { FormEvent, useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
-import { CardCompanyType, CardInfo, PageInfo } from '../../types';
+import { CardNumberInput } from '../../components/addCardPage/CardNumberInput';
+import { CvcInput } from '../../components/addCardPage/CvcInput';
+import { ExpirationInput } from '../../components/addCardPage/ExpirationInput';
+import { OwnerInput } from '../../components/addCardPage/OwnerInput';
+import { PasswordInput } from '../../components/addCardPage/PasswordInput';
+import { RegisteredCard } from '../../components/addCardPage/RegisteredCard';
+import { SelectCardCompany } from '../../components/addCardPage/SelectCardCompany';
+import { Button } from '../../components/common/Button';
+import { Card } from '../../components/common/Card';
+import { ChangeButton } from '../../components/common/ChangeButton';
+import { InputField } from '../../components/common/InputField';
+import { Modal } from '../../components/common/Modal';
+import { PrevButton } from '../../components/common/PrevButton';
+import { CARD_COMPANY_DATA } from '../../constants/constant';
+import { InputValuesInformationProps } from '../../hooks/createFormInputValue';
+import { getFormValidateResult } from '../../hooks/getFormValidateResult';
 import { useFocusInput } from '../../hooks/useFocusInput';
 import { useFormInputs } from '../../hooks/useFormInputs';
 import { useHideScrollState } from '../../hooks/useHideScrollState';
+import { CardCompanyType, CardInfo, PageInfo } from '../../types/types';
 import {
   createUniqueId,
   isPrevDate,
   setNextInputFocus,
   userConfirm,
 } from '../../utils';
-import { InputValuesInformationProps } from '../../hooks/createFormInputValue';
-import { getFormValidateResult } from '../../hooks/getFormValidateResult';
-import { RegisteredCard } from '../../components/addCardPage/RegisteredCard';
-import { PrevButton } from '../../components/common/PrevButton';
-import { Card } from '../../components/common/Card';
-import { ChangeButton } from '../../components/common/ChangeButton';
-import { InputField } from '../../components/common/InputField';
-import { CardNumberInput } from '../../components/addCardPage/CardNumberInput';
-import { ExpirationInput } from '../../components/addCardPage/ExpirationInput';
-import { OwnerInput } from '../../components/addCardPage/OwnerInput';
-import { CvcInput } from '../../components/addCardPage/CvcInput';
-import { Button } from '../../components/common/Button';
-import { CARD_COMPANY_DATA } from '../../constant';
-import { SelectCardCompany } from '../../components/addCardPage/SelectCardCompany';
-import { Modal } from '../../components/common/Modal';
-import { PasswordInput } from '../../components/addCardPage/PasswordInput';
 
 interface AddCardPageProps {
   cardList: CardInfo[];
